% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.bsem}
\alias{plot.bsem}
\title{'bsem' object plot}
\usage{
\method{plot}{bsem}(x, digits = 2, fontsize = 15, width = 2, size = 10, ...)
}
\arguments{
\item{x}{an object of class bsem}

\item{digits}{number of digits to display}

\item{fontsize}{edge fontsize (estimates character size)}

\item{width}{lines width}

\item{size}{nodes size (for both: ellipses and boxes)}

\item{...}{further arguments passed to \code{visNetwork::visNetwork} function}
}
\value{
a visNetwork graph:
\describe{
   \item{}{The ellipsoidal nodes represent the latent variables}
   \item{}{The boxes represent the manifest variables}
   \item{}{The dashed lines represent the linear relations between latent scores}
   \item{}{The solid lines represent the relationship between the manifest an latent variables}
   \item{}{The recursive solid lines refers to the error variance estimate of each manifest variable}
}
}
\description{
'bsem' object plot
}
