% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bsem.R
\name{print.bsem}
\alias{print.bsem}
\title{'bsem' object print}
\usage{
\method{print}{bsem}(x, digits = max(getOption("digits") - 4, 3), ...)
}
\arguments{
\item{x}{an object of class spbp}

\item{digits}{number of digits to display}

\item{...}{further arguments passed to or from print methods}
}
\value{
none
}
\description{
'bsem' object print
}
