% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{Run a shiny app}
\usage{
runShiny()
}
\description{
An introduction to the bsem package
 @details
 \describe{
   \item{signals}{this option is not allowed, it is only available in \code{\link[bsem]{sem}}}
   \item{stanfit}{S4 object of class stanfit}
 }
}
\examples{

\dontrun{
library("bsem")
runShiny()
}
}
\seealso{
\code{\link[bsem]{sem}}, \code{\link[bsem]{simdata}}, \code{\link[bsem]{arrayplot}}, \code{\link[bsem]{summary.bsem}}, \code{\link[bsem]{print.bsem}}
}
\author{
RV Panaro
}
