% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{sem}
\alias{sem}
\title{sem: The SEM Function}
\usage{
sem(
  data,
  paths,
  blocks,
  signals,
  row_names = rownames(data),
  prior_specs = list(coef = c("normal(0,1)"), error_var = c("inv_gamma(2.1, 1.1)")),
  cores = parallel::detectCores(),
  pars = c("alpha", "lambda", "sigma2", "Xna"),
  iter = 2000,
  chains = 4,
  scaled = TRUE,
  ...
)
}
\arguments{
\item{data}{a mandatory data.frame object where the columns are variables and the rows are observations}

\item{paths}{list referring to the inner model}

\item{blocks}{list of column names (or integers in 1:ncol(data)) indicating the manisfest variables correpoding to each block;}

\item{signals}{list referring to the signals of the factor loadings initial values; must be true: (length(signals) == length(blocks)) && (lengths(signals) == lengths(blocks))}

\item{prior_specs}{prior settings for the Bayesian approach; only `normal(0,...)` for coef and `inv_gamma` for error_var are temporarily available, beta is ignored if paths are not specific; next versions should include more options}

\item{cores}{number of core threads to be used}

\item{pars}{which parameters to be included in the outcome; options are any subset of default c("alpha", "lambda", "sigma2", "Xna")}

\item{iter}{number of iterations}

\item{chains}{number of posterior chains}

\item{scaled}{logical; indicates whether to center and scale the data}

\item{...}{further arguments passed to or from other methods}

\item{rownames}{refers to the observation vector of characters; must have nrow(data) lenght; defaults to rownames(data)}
}
\value{
An object of class \code{sem}; a list of 14 to 16:
\describe{
   \item{stanfit}{S4 object of class stanfit}
   \item{posterior}{the list of posterior draws separate by chains}
   \item{model}{character; pointer to pre-defined stan model}
   \item{mean_loadings}{matrix of factor loadings posterior means}
   \item{mean_scores}{matrix of factor scores posterior means}
   \item{mean_var}{vector of error variances posterior means}
   \item{coef}{vector of regression coefficients posterior means}
   \item{stats}{posterior descriptives statistics}
   \item{blocks}{list of blocks}
   \item{paths}{list of paths}
   \item{credint}{Highest posterior density intervals (HPD)}
   \item{h}{vector of communalities}
   \item{PVTE}{vector of total variance proportions}
   \item{AFR2}{adjusted coefficient of determination}
   \item{SQE}{explained sums of squares}
   \item{SQT}{total sums of squares}
}
}
\description{
Structural Equation Models (SEM) and particular cases via rstan interface
}
\details{
Fits the SEM to specific data
}
\examples{
data('set1')

fit1 <- sem(data = set1$set, blocks = set1$blocks,
 chains = 1)
summary(fit1)

\dontrun{
data('set2')

fit2 <- sem(data = set2$set, blocks = set2$blocks,
 chains = 4, iter = 2500, warmup = 1000)
summary(fit2)
}

}
