% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\docType{data}
\name{set1}
\alias{set1}
\title{Data set 1; generated from a factor analysis model}
\format{
A list of 3 objects contaning the data set, the real (true) values and the blocks of variables that manifest the latent scores
\describe{
   \item{set}{the artificial data set; \code{double[20x100]} }
   \item{real}{a list of 3 containing the true values of the parameters; alpha = loadings, lambda = scores, sigma2 = error variance}
   \item{blocks}{list of 5 integer arrays describing the relationship between observed and unobserved variables}
}
}
\usage{
data(set1)
}
\description{
Data set 1; generated from a factor analysis model
}
\examples{
data(set1)
}
\keyword{datasets}
\keyword{set1}
\keyword{set2}
