% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayplot.R
\name{arrayplot}
\alias{arrayplot}
\title{Array Plotting}
\usage{
arrayplot(
  y,
  mini = -max(abs(min(y)), max(y)),
  maxi = max(abs(min(y)), max(y)),
  colors = 1,
  ylab = "",
  xlab = "",
  main = ""
)
}
\arguments{
\item{mini, maxi}{color range.}

\item{colors}{1 = blue/red, 0 = white/black.}

\item{matrix}{a numerical data.frame or matrix object.}
}
\description{
Graphical representation for matrix class objects using lattice package.
}
\examples{

data(iris)

plotMatrix(iris[, c(1,2,3)])
plotMatrix(iris[, c(1,2,3)], colors = 0)
}
