% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{sem}
\alias{sem}
\title{sem: The SEM Function}
\usage{
sem(
  data,
  paths,
  blocks,
  signals,
  row_names,
  priors = list(beta = c("normal(0,1)"), noise = c("gamma(2.1, 1.1)")),
  cores = parallel::detectCores(),
  pars = c("alpha", "lambda", "sigma2", "Xna", "log_lik"),
  iter = 4000,
  chains = 4,
  scaled = TRUE,
  ...
)
}
\arguments{
\item{data}{a mandatory data.frame object where the columns are variables and the rows are observations}

\item{paths}{list referring to the inner model}

\item{blocks}{list of column names indicating the manisfest variables correpoding to each block}

\item{priors}{prior settings for the Bayesian approach; `normal` or `cauchy` for beta; `gamma`, `inv_gamma` or `lognormal` for gamma (BP coefficients)}

\item{cores}{number of core threads to use}

\item{scaled}{logical; indicates whether to center and scale the data}

\item{...}{further arguments passed to or from other methods}
}
\value{
An object of class \code{sem}
}
\description{
Structural Equation Models (SEM) and particular cases via rstan interface
}
\details{
Fits the SEM to specific data
}
