% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bsem.R
\name{summary.bsem}
\alias{summary.bsem}
\title{'bsem' object summary}
\usage{
\method{summary}{bsem}(object, digits = 4, ...)
}
\arguments{
\item{object}{an object of class spbp}

\item{digits}{number of digits to display}

\item{...}{further arguments passed to or from summary methods}
}
\value{
none
}
\description{
'bsem' object summary
}
